refactor: extract HTML report into scripts/alert_report.py
Move read_code_snippet, render_html, and write_summary out of
post_alert_action.py into a dedicated module. Keeps the action
script focused on orchestration.
